check file exist

+++ app/Http/Controllers/Admin/AdminController.php
@@ -47,7 +47,10 @@ class AdminController extends Controller
     	$splitPath = explode("/", $url["path"]);
         $splitPathLength = count($splitPath);
         $filename=$splitPath[$splitPathLength-1];
-        unlink(public_path('images/froalafiles/'.$filename));
+        if(file_exists(public_path('images/froalafiles/'.$filename)))
+        {
+            unlink(public_path('images/froalafiles/'.$filename));
+        }
         FroalaFileUpload::where('path', 'LIKE', '%' . $filename . '%')->delete();
     }
 }
